feat: adds argparse and url request functions
* Add argparse

* URL request functions

* Additional code modularization

* Type checks for new and existing functions

* Refactor urban to _urban

Issue: #55
Tag: 2.0.0

Authored-by: Joshua Rose <joshuarose099@gmail.com>
