+++ src/urban_api.py
@@ -3,11 +3,10 @@
 API
 ===
 
-:Authors:
-    Joshua Rose
-
+:Author: Joshua Rose
 :Version: 2.0.0 of 2023/05/28
 :License: `Apache 2.0 <https://gh-syn.github.io/urban-cli/license.html>`_
+:File: urban_api.py
 
 The API contains centralized interactions with urban-cli.
 It's responsible for the following functions.
@@ -16,7 +15,7 @@ Sending requests
 ----------------
 Requests that need to be sent to a remote source will be done so through
 functions declared in this file. Most of which, will be using custom
-exception classes as normal exceptions will not suit this specific use case..
+exception classes as normal exceptions will not suit this specific use case.
 
 Processing responses
 --------------------
@@ -31,7 +30,10 @@ within the API file. This file serves as a central point from which functions ar
 accessed.
 """
 
+from typing import Literal
+
 import requests
+import rich
 
 from src.urban_exceptions import InvalidStatusCodeError
 
@@ -50,34 +52,66 @@ def apply_word_to_url(word: str) -> str:
     return fetch_url
 
 
-def send_exists_request(word: str) -> bool:
+def send_exists_request(word: str) -> Literal[False] | requests.Response:
     """
     Requests if a word exists from the urban dictionary.
 
     :param word: Word attaches as postfix when requesting exists.
     :raise InvalidStatusCodeError: If response status code does not equal 200 or 404.
-    :return: `True` if `word` exists otherwise returns `False`.
+    :return: If the word exists or if the word doesn't exist. Return types differ.
+    :rtype `requests.Response`: if `word` exists in the Urban Dictionary.
+    :rtype `False`: If `word` does not exist in the Dictionary.
     """
 
     # Attach word to url variable
     url = apply_word_to_url(word)
 
+    # Get reponse from urban dictionary as a `requests.Response`.
     response = requests.get(url)
     status = response.status_code
 
     match status:
         case 200:
-            return True
+            return response
         case 404:
             return False
         case _:
             raise InvalidStatusCodeError(status)
 
 
-def send_phrase_request(phrase: str):
+def send_phrase_request(phrase: str) -> bytes:
     """
     Request phrase from the urban dictionary.
 
+    Handles the case that the phrase does not exist with an error message.
+    Otherwise, it returns the received phrase that it got from `send_exists_request()` function.
+
     :param phrase: Queries to URL in `requests.Request` object.
-    :return:
+    :raise TypeError: If `phrase` is not a `string`.
+    :raise SystemExit: If `phrase` returns a 404. That is - the phrase doesn't exist.
+    :return: Received data from the `requests.Response` content method.
+    :rtype: bytes
     """
+
+    if not isinstance(phrase, str):
+        raise TypeError(f"`phrase` read as a `{type(phrase)}` must be a `string`.")
+
+    # TODO: format `phrase` with urllib before asking if it exists
+    phrase_exists = send_exists_request(phrase)
+
+    if not phrase_exists:
+        # `phrase` doesn't exist in the dictionary ... !
+        rich.print("The word {} does'nt exist yet!", end="")
+        rich.print(
+            "You can change that by submitting a definition on the {link}".format(
+                link="[link url='https://urbandictionary.com/']Urban Dictionary website.[/link]"
+            )
+        )
+
+        raise SystemExit
+
+    # We now know that the phrase not only exists, but is a response.
+    phrase_response = phrase_exists.content
+
+    # We return the response content
+    return phrase_response
