+++ src/urban_utils.py
@@ -1,15 +1,29 @@
 """
-### Meta & Licensing
-__File__: `urban_utils.py`
-__License__: [Apache 2.0](https://gh-syn.github.io/urban-cli/license.html)
-__Author__: Joshua Rose <joshuarose099@gmail.com>
+=====
+Utils
+=====
 
-### Module Description
+:Author: Joshua Rose
+:Version: 2.0.0 of 2023/05/28
+:License: `Apache 2.0 <https://gh-syn.github.io/urban-cli/license.html>`_
+:File: urban_utils.py
 
-Contains commonly used utility functions.
+This file contains utility functions and helper methods.
+It is responsible for the following functions.
 
-This module handles:
- - common data structures
- - data manipulation
- - io handling
+Data Manipulation
+-----------------
+When a request object is received, it needs to be formatted in a way that Python can understand
+The beautifulsoup library helps beautifully ✨ with this. It parses html data in a way that can be 
+read and manipulated very easily.
+
+I/O handling
+------------
+The argparse library can be thought of as boilerplate level code in addition to a few extra classes 
+that greatly help building a command line interface. It removes most of the argument logic and lets the
+developer focus on what's important. Argparse is used in this file for showing and receiving data.
+
+Common Data Structures
+-----------------------
+Common data structures such as a date, definition struct, lives in this file.
 """
